refactor(cli): extract config, auth, models, and automation groups
Move config (show/edit/get/set/keys), auth (vault, `auth chatgpt`,
credential pools, hash-password), models (list/aliases/providers/set +
approvals), and automation (workflow + trigger + cron) handlers out of
main.rs. DeviceAuthNextStep moves with the auth group.

The auth group is named `auth` (not `secrets`) so the source file does
not trip the repo's sensitive-filename git guard.

Pure code move, no behavior change.
